Refactor/#17/InputField 관련 코드 수정
